Made a smart enemy
Made an enemy that follows player

+++ src/com/survivor/main/HUD.java
@@ -11,7 +11,7 @@ public class HUD {
 	private int level = 1;
 	
 	public void tick(){
-		HEALTH = Game.clamp(HEALTH, 0, 100);
+		HEALTH = (int) Game.clamp(HEALTH, 0, 100);
 		
 		score ++;
 	}
